// src/models/garmentModel.ts
export type garmentModel = {
    name: string;
    token: number;
    color: string;
    type: string;
    originalPricing: string;
    season: string;
    fibers: string;
    coating: string;
    textiles: string;
    factory: string;
    manufactureDate: string;
    designDate: string;
    size: string;
    damageSustained?: string;
    lifeSpan?: string;
    currentEnvironment?: string;
    returned?: boolean;
    dateOfSale?: string;
    salePrice?: string;
    shippingDate?: string;
    shipper?: string;
    warrantyTime?: string;
    imageCID: string;
}